requirements.txt
fastapi
uvicorn[standard]
aiomysql
motor
boto3
cryptography
httpx
python-dotenv
pydantic-settings
openai
beautifulsoup4
pytest
BeautifulSoup4
konlpy
transformers
tf-keras
torch
nltk
git+https://github.com/haven-jeon/PyKoSpacing.git#egg=pykospacing
git+https://github.com/ssut/py-hanspell.git#egg=py-hanspell